Move packet and akcnowledgement processing pipelines to core-protocol
