+++ packages/core/crates/core-protocol/src/msg/mod.rs
@@ -1,1 +1,2 @@
 pub mod config;
+pub mod processor;
